refactor: split client.ts into focused modules
Extract 1658-line monolith into 4 focused modules:
- client.ts (573): class shell, lifecycle, RPC transport, delegation
- api.ts (647): stateless low-level API operations (RpcCaller pattern)
- workflows.ts (473): run* orchestrations + polling helpers
- download.ts (378): file download, cookie jar, type-specific saves

Public API unchanged — NotebookClient retains all methods as thin wrappers.

Closes #8
